+++ lib/src/audio_isolate.dart
@@ -285,7 +285,7 @@ void audioIsolate(SendPort isolateToMainStream) {
         for (final sound in activeSounds) {
           isolateToMainStream.send(
             (
-              event: SoundEvent.soundDisposed,
+              event: SoundEventType.soundDisposed,
               sound: sound,
               handle: 0,
             ),
@@ -390,7 +390,7 @@ void audioIsolate(SendPort isolateToMainStream) {
 
                   isolateToMainStream.send(
                     (
-                      event: SoundEvent.handleIsNoMoreValid,
+                      event: SoundEventType.handleIsNoMoreValid,
                       sound: sound,
                       handle: handle,
                     ),