+++ test_fixes/playererrors_multipleinitialization.dart
@@ -1,3 +1,0 @@
-import 'package:flutter_soloud/flutter_soloud.dart';
-
-PlayerErrors error = PlayerErrors.isolateAlreadyStarted;