+++ CHANGELOG.md
@@ -30,7 +30,17 @@ to `2.0.0-pre.2` and beyond.
     print('Oh no! $e');
   }
   ```
-
+- All time-related parameters and return values are now `Duration` type.
+  Before, they were `double`.
+- Fixed velocity computation bug in `example/`.
+- Renamed `SoundEvent` to `SoundEventType`. Quick fix available.
+- `SoundProps.soundEvents` is now a `Stream`, not a `StreamController`
+- `SoundProps.soundEvents` stream is now closed automatically when
+  `SoLoud.disposeSound()` is called.
+- `SoLoud.activeSounds` is now an `Iterable` instead of a `List`
+  (therefore, it cannot be modified from outside the package).
+- Renamed `SoLoud.getFxParams` to `SoLoud.getFilterParams`. Quick fix available.
+- Renamed `SoLoud.setFxParams` to `SoLoud.setFilterParams`. Quick fix available.
 
 #### 2.0.0-pre.1 (12 Mar 2024)
 - added `looping` and `loopingStartAt` properties to `SoLoud.play()` and `SoLoud.play3d()`.